network manager is converted to a singleton class
